requirements/base.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autoflake==2.2.0
billiard==4.2.1
black==23.7.0
blinker==1.9.0
celery==5.3.6
cffi==1.17.1
cfgv==3.4.0
classify-imports==4.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
distlib==0.3.9
dj-database-url==2.2.0
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
django-jazzmin==3.0.1
django-modeltranslation==0.19.14
django-recaptcha==3.0.0
django-redis==5.4.0
django-resized==1.0.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
filelock==3.18.0
flake8==7.2.0
flower==2.0.1
humanize==4.12.3
identify==2.6.10
inflection==0.5.1
isort==5.12.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
nplusone==1.0.0
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pre_commit==4.2.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pyflakes==3.3.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-environ==0.4.54
pytz==2025.2
pyupgrade==3.10.1
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
reorder-python-imports==3.10.0
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
tokenize_rt==6.1.0
tomli==2.2.1
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
vine==5.1.0
virtualenv==20.31.1
wcwidth==0.2.13
yesqa==1.5.0
